fix: choose games randomly
