+++ src/main/java/com/shanebeestudios/skbee/Main.java
@@ -1,24 +1,21 @@
 package com.shanebeestudios.skbee;
 
+import com.shanebeestudios.skbee.checkers.Checkers;
 import com.shanebeestudios.skbee.pong.Pong;
 
 public class Main {
     public static void main(String[] args) {
         var games = new Class<?>[] {
                 Pong.class,
-                com.shanebeestudios.skbee.checkers.Main.class
+                Checkers.class
         };
 
-        // for now
-        new com.shanebeestudios.skbee.checkers.Main();
-        return;
-
-//        try {
-//            var game = games[(int) Math.floor(Math.random() * games.length)];
-//            game.getConstructor().newInstance();
-//        } catch (ReflectiveOperationException ignored) {
-//            System.out.println("we failed to pick a game :( defaulting to pong");
-//            new Pong();
-//        }
+        try {
+            var game = games[(int) Math.floor(Math.random() * games.length)];
+            game.getConstructor().newInstance();
+        } catch (ReflectiveOperationException ignored) {
+            System.out.println("we failed to pick a game :( defaulting to pong");
+            new Pong();
+        }
     }
 }
